+++ tests/test_templates/test_base_html.py
@@ -3,7 +3,7 @@ from jinja2 import DictLoader
 
 from render_engine.page import Page
 from render_engine.site import Site
-from render_engine.utils.themes import Theme
+from render_engine.themes import Theme
 
 
 @pytest.fixture()
