+++ tests/test_theme_manager.py
@@ -1,7 +1,7 @@
 from jinja2.environment import Environment
 from jinja2.loaders import ChoiceLoader, DictLoader
 
-from render_engine.utils.themes import Theme, ThemeManager
+from render_engine.themes import Theme, ThemeManager
 
 
 def test_ThemeManager_builds():
